refactor(plugins): reorganize plugin structure to make browsing easier
