+++ lua/astronvim/plugins/_astrotheme.lua
@@ -1,0 +1,1 @@
+return { "AstroNvim/astrotheme", lazy = true, opts = { plugins = { ["dashboard-nvim"] = true } } }
