+++ lua/astronvim/plugins/autopairs.lua
@@ -1,0 +1,29 @@
+return {
+  "windwp/nvim-autopairs",
+  event = "User AstroFile",
+  dependencies = {
+    {
+      "AstroNvim/astrocore",
+      opts = function(_, opts)
+        local maps = opts.mappings
+        maps.n["<Leader>ua"] = { function() require("astrocore.toggles").autopairs() end, desc = "Toggle autopairs" }
+      end,
+    },
+  },
+  opts = {
+    check_ts = true,
+    ts_config = { java = false },
+    fast_wrap = {
+      map = "<M-e>",
+      chars = { "{", "[", "(", '"', "'" },
+      pattern = ([[ [%'%"%)%>%]%)%}%,] ]]):gsub("%s+", ""),
+      offset = 0,
+      end_key = "$",
+      keys = "qwertyuiopzxcvbnmasdfghjkl",
+      check_comma = true,
+      highlight = "PmenuSel",
+      highlight_grey = "LineNr",
+    },
+  },
+  config = function(...) require "astronvim.plugins.configs.nvim-autopairs"(...) end,
+}
