+++ lua/astronvim/plugins/better-escape.lua
@@ -1,0 +1,1 @@
+return { "max397574/better-escape.nvim", event = "InsertCharPre", opts = { timeout = 300 } }
