+++ lua/astronvim/plugins/cmp_luasnip.lua
@@ -43,6 +43,9 @@ return {
         end,
         preselect = cmp.PreselectMode.None,
         formatting = { fields = { "kind", "abbr", "menu" } },
+        completion = {
+          completeopt = "menu,menuone,noinsert",
+        },
         confirm_opts = {
           behavior = cmp.ConfirmBehavior.Replace,
           select = false,
@@ -148,42 +151,5 @@ return {
       region_check_events = "CursorMoved",
     },
     config = function(...) require "astronvim.plugins.configs.luasnip"(...) end,
-  },
-  {
-    "onsails/lspkind.nvim",
-    lazy = true,
-    enabled = vim.g.icons_enabled ~= false,
-    dependencies = {
-      {
-        "hrsh7th/nvim-cmp",
-        opts = function(_, opts)
-          if not opts.formatting then opts.formatting = {} end
-          opts.formatting.format = require("lspkind").cmp_format(require("astrocore").plugin_opts "lspkind.nvim")
-        end,
-      },
-    },
-    opts = {
-      mode = "symbol",
-      symbol_map = {
-        Array = "󰅪",
-        Boolean = "⊨",
-        Class = "󰌗",
-        Constructor = "",
-        Key = "󰌆",
-        Namespace = "󰅪",
-        Null = "NULL",
-        Number = "#",
-        Object = "󰀚",
-        Package = "󰏗",
-        Property = "",
-        Reference = "",
-        Snippet = "",
-        String = "󰀬",
-        TypeParameter = "󰊄",
-        Unit = "",
-      },
-      menu = {},
-    },
-    config = function(...) require "astronvim.plugins.configs.lspkind"(...) end,
   },
 }
