+++ lua/astronvim/plugins/colorizer.lua
@@ -1,0 +1,15 @@
+return {
+  "NvChad/nvim-colorizer.lua",
+  dependencies = {
+    {
+      "AstroNvim/astrocore",
+      opts = function(_, opts)
+        local maps = opts.mappings
+        maps.n["<Leader>uz"] = { "<Cmd>ColorizerToggle<CR>", desc = "Toggle color highlight" }
+      end,
+    },
+  },
+  event = "User AstroFile",
+  cmd = { "ColorizerToggle", "ColorizerAttachToBuffer", "ColorizerDetachFromBuffer", "ColorizerReloadAllBuffers" },
+  opts = { user_default_options = { names = false } },
+}
