+++ lua/astronvim/plugins/comment.lua
@@ -1,0 +1,28 @@
+return {
+  "numToStr/Comment.nvim",
+  dependencies = {
+    { "JoosepAlviste/nvim-ts-context-commentstring", opts = { enable_autocmd = false } },
+    {
+      "AstroNvim/astrocore",
+      opts = function(_, opts)
+        local maps = opts.mappings
+        maps.n["<Leader>/"] = {
+          function() require("Comment.api").toggle.linewise.count(vim.v.count > 0 and vim.v.count or 1) end,
+          desc = "Toggle comment line",
+        }
+        maps.v["<Leader>/"] = {
+          "<Esc><Cmd>lua require('Comment.api').toggle.linewise(vim.fn.visualmode())<CR>",
+          desc = "Toggle comment for selection",
+        }
+      end,
+    },
+  },
+  keys = {
+    { "gc", mode = { "n", "v" }, desc = "Comment toggle linewise" },
+    { "gb", mode = { "n", "v" }, desc = "Comment toggle blockwise" },
+  },
+  opts = function(_, opts)
+    local commentstring_avail, commentstring = pcall(require, "ts_context_commentstring.integrations.comment_nvim")
+    if commentstring_avail then opts.pre_hook = commentstring.create_pre_hook() end
+  end,
+}
