+++ lua/astronvim/plugins/dressing.lua
@@ -1,0 +1,9 @@
+return {
+  "stevearc/dressing.nvim",
+  lazy = true,
+  init = function() require("astrocore").load_plugin_with_func("dressing.nvim", vim.ui, "input", "select") end,
+  opts = {
+    input = { default_prompt = "➤ " },
+    select = { backend = { "telescope", "builtin" } },
+  },
+}
