+++ lua/astronvim/plugins/guess-indent.lua
@@ -1,0 +1,5 @@
+return {
+  "NMAC427/guess-indent.nvim",
+  event = "User AstroFile",
+  config = function(...) require "astronvim.plugins.configs.guess-indent"(...) end,
+}
