+++ lua/astronvim/plugins/neo-tree.lua
@@ -1,6 +1,7 @@
 return {
   "nvim-neo-tree/neo-tree.nvim",
   dependencies = {
+    "nvim-lua/plenary.nvim",
     "MunifTanjim/nui.nvim",
     {
       "AstroNvim/astrocore",
