+++ lua/astronvim/plugins/toggleterm.lua
@@ -1,0 +1,70 @@
+return {
+  "akinsho/toggleterm.nvim",
+  cmd = { "ToggleTerm", "TermExec" },
+  dependencies = {
+    {
+      "AstroNvim/astrocore",
+      opts = function(_, opts)
+        local maps = opts.mappings
+        local astro = require "astrocore"
+        maps.n["<Leader>t"] = vim.tbl_get(opts, "_map_sections", "t")
+        if vim.fn.executable "lazygit" == 1 then
+          maps.n["<Leader>g"] = vim.tbl_get(opts, "_map_sections", "g")
+          maps.n["<Leader>gg"] = {
+            function()
+              local worktree = astro.file_worktree()
+              local flags = worktree and (" --work-tree=%s --git-dir=%s"):format(worktree.toplevel, worktree.gitdir)
+                or ""
+              astro.toggle_term_cmd("lazygit " .. flags)
+            end,
+            desc = "ToggleTerm lazygit",
+          }
+          maps.n["<Leader>tl"] = maps.n["<Leader>gg"]
+        end
+        if vim.fn.executable "node" == 1 then
+          maps.n["<Leader>tn"] = { function() astro.toggle_term_cmd "node" end, desc = "ToggleTerm node" }
+        end
+        local gdu = vim.fn.has "mac" == 1 and "gdu-go" or "gdu"
+        if vim.fn.executable(gdu) == 1 then
+          maps.n["<Leader>tu"] = { function() astro.toggle_term_cmd(gdu) end, desc = "ToggleTerm gdu" }
+        end
+        if vim.fn.executable "btm" == 1 then
+          maps.n["<Leader>tt"] = { function() astro.toggle_term_cmd "btm" end, desc = "ToggleTerm btm" }
+        end
+        local python = vim.fn.executable "python" == 1 and "python" or vim.fn.executable "python3" == 1 and "python3"
+        if python then
+          maps.n["<Leader>tp"] = { function() astro.toggle_term_cmd(python) end, desc = "ToggleTerm python" }
+        end
+        maps.n["<Leader>tf"] = { "<Cmd>ToggleTerm direction=float<CR>", desc = "ToggleTerm float" }
+        maps.n["<Leader>th"] =
+          { "<Cmd>ToggleTerm size=10 direction=horizontal<CR>", desc = "ToggleTerm horizontal split" }
+        maps.n["<Leader>tv"] = { "<Cmd>ToggleTerm size=80 direction=vertical<CR>", desc = "ToggleTerm vertical split" }
+        maps.n["<F7>"] = { "<Cmd>ToggleTerm<CR>", desc = "Toggle terminal" }
+        maps.t["<F7>"] = maps.n["<F7>"]
+        maps.n["<C-'>"] = maps.n["<F7>"] -- requires terminal that supports binding <C-'>
+        maps.t["<C-'>"] = maps.n["<F7>"] -- requires terminal that supports binding <C-'>
+      end,
+    },
+  },
+  opts = {
+    highlights = {
+      Normal = { link = "Normal" },
+      NormalNC = { link = "NormalNC" },
+      NormalFloat = { link = "NormalFloat" },
+      FloatBorder = { link = "FloatBorder" },
+      StatusLine = { link = "StatusLine" },
+      StatusLineNC = { link = "StatusLineNC" },
+      WinBar = { link = "WinBar" },
+      WinBarNC = { link = "WinBarNC" },
+    },
+    size = 10,
+    on_create = function()
+      vim.opt.foldcolumn = "0"
+      vim.opt.signcolumn = "no"
+    end,
+    open_mapping = [[<F7>]],
+    shading_factor = 2,
+    direction = "float",
+    float_opts = { border = "rounded" },
+  },
+}
