+++ lua/astronvim/plugins/ts-autotag.lua
@@ -1,0 +1,1 @@
+return { "windwp/nvim-ts-autotag", event = "User AstroFile", opts = {} }
