+++ lua/astronvim/plugins/which-key.lua
@@ -1,0 +1,8 @@
+return {
+  "folke/which-key.nvim",
+  event = "VeryLazy",
+  opts = {
+    icons = { group = vim.g.icons_enabled ~= false and "" or "+", separator = "" },
+    disable = { filetypes = { "TelescopePrompt" } },
+  },
+}
