+++ lua/astronvim/plugins/window-picker.lua
@@ -1,0 +1,6 @@
+return {
+  "s1n7ax/nvim-window-picker",
+  main = "window-picker",
+  lazy = true,
+  opts = { picker_config = { statusline_winbar_picker = { use_winbar = "smart" } } },
+}
